Changed text-formatting functions from infodesk file to the page file
