+++ app_feup/lib/view/Widgets/info_desk_card.dart
@@ -1,43 +1,6 @@
 import 'package:flutter/material.dart';
 import 'generic_card.dart';
-
-Container h1(String text, BuildContext context, {bool initial = false}) {
-  final double marginTop = initial ? 15.0 : 30.0;
-  return Container(
-      margin: EdgeInsets.only(top: marginTop, bottom: 0.0, left: 20.0),
-      child: Align(
-        alignment: Alignment.centerLeft,
-        child: Text(text,
-            style:
-                Theme.of(context).textTheme.headline4.apply(fontSizeDelta: 0)),
-      ));
-}
-
-Container h2(String text, BuildContext context) {
-  return Container(
-      margin: const EdgeInsets.only(top: 13.0, bottom: 0.0, left: 20.0),
-      child: Align(
-        alignment: Alignment.centerLeft,
-        child: Text(text,
-            style:
-                Theme.of(context).textTheme.headline4.apply(fontSizeDelta: -4)),
-      ));
-}
-
-Container infoText(String text, BuildContext context, {bool last = false}) {
-  final double marginBottom = last ? 8.0 : 0.0;
-  return Container(
-      margin: EdgeInsets.only(top: 8, bottom: marginBottom, left: 20.0),
-      child: Align(
-        alignment: Alignment.centerLeft,
-        child: InkWell(
-          child: Text(
-            text,
-            style: Theme.of(context).textTheme.headline3,
-          ),
-        ),
-      ));
-}
+import '../Pages/useful_contacts_card_page_view.dart';
 
 class InfoDeskCard extends GenericCard {
   InfoDeskCard({Key key}) : super(key: key);
